requirements.txt
# Web Test Automation Stack
# Core browser automation
playwright==1.41.0
browser-use==0.1.5

# AI/LLM integration
crewai==0.28.8
openai==1.12.0

# Data handling and validation
pydantic==2.6.1
pyyaml==6.0.1

# Environment and configuration
python-dotenv==1.0.0

# Logging and observability
structlog==23.2.0

# Testing utilities
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-playwright==0.4.4

# Development tools
black==24.1.1
flake8==7.0.0
mypy==1.8.0 